Add carousel that shows 5 slides and does not change the width of the cards
This required a combination of using the badly named `variableWidth: true` option, that prevents slick from dynamically calculating the slide width, and setting the correct `width` on the container, so that the 5 slides together with their margins fit into the container exactly. If the container is wider than needed for displaying `slidesToShow` slides, more slides will be displayed.